% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy1.R
\name{fuzzy1}
\alias{fuzzy1}
\title{adding normal variables to a vector}
\usage{
fuzzy1(x)
}
\arguments{
\item{an}{input vector/number}
}
\value{
the input \code{x} + \code{rnorm}
}
\description{
adding normal variables to a vector
}
\examples{
 fuzzy1(1:5)
}
