% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy2.R
\name{fuzzy2}
\alias{fuzzy2}
\title{call the clean1 subroutine}
\usage{
fuzzy2(n)
}
\arguments{
\item{n}{an integer length}

\item{x}{vector of length n}
}
\value{
\code{y(n)}
}
\description{
call the clean1 subroutine
}
